add donations and activity log modules and fix issues

+++ app/Http/Controllers/BaseController.php
@@ -20,25 +20,30 @@ abstract class BaseController extends Controller
 
     public function index()
     {
+        // $this->authorize('view', $this->model);
         return view($this->viewPath . '.index');
     }
 
     public function data()
     {
+        // $this->authorize('view', $this->model);
         $query = $this->model::query();
         return $this->getDataTable($query);
     }
 
     public function create()
     {
+        $this->authorize('create', $this->model);
         return view($this->viewPath . '.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', $this->model);
+
         try {
             $data = $request->validate($this->validationRules);
-            if($data['date']){
+            if(isset($data['date'])){
                 $convertedDateTime = str_replace('T', ' ', $data['date']) . ':00';
                 $data['date'] = $convertedDateTime;
             }
@@ -47,7 +52,7 @@ abstract class BaseController extends Controller
             if ($request->ajax()) {
                 return response()->json([
                     'success' => true,
-                    'message' => class_basename($this->model) . ' created successfully',
+                    'message' => __('messages.' . class_basename($this->model) . ' created successfully'),
                     'data' => $item
                 ]);
             }
@@ -58,23 +63,27 @@ abstract class BaseController extends Controller
             if ($request->ajax()) {
                 return response()->json([
                     'success' => false,
-                    'message' => 'Error creating ' . class_basename($this->model),
+                    'message' => __('messages.' . class_basename($this->model) . ' created failed'),
                     'errors' => $e->getMessage()
                 ], 422);
             }
 
-            return back()->withErrors(['error' => 'Error creating ' . class_basename($this->model)])->withInput();
+            return back()->withErrors(['error' => __('messages.' . class_basename($this->model) . ' created failed')])->withInput();
         }
     }
 
     public function edit($id)
     {
+        $this->authorize('update', $this->model);
+
         $item = $this->model::findOrFail($id);
         return view($this->viewPath . '.edit', compact('item'));
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('update', $this->model);
+
         try {
             $item = $this->model::findOrFail($id);
             
@@ -87,28 +96,30 @@ abstract class BaseController extends Controller
             if ($request->ajax()) {
                 return response()->json([
                     'success' => true,
-                    'message' => class_basename($this->model) . ' updated successfully',
+                    'message' => __('messages.' . class_basename($this->model) . ' updated successfully'),
                     'data' => $item
                 ]);
             }
 
             return redirect()->route($this->routePrefix . '.index')
-                ->with('success', class_basename($this->model) . ' updated successfully');
+                ->with('success', __('messages.' . class_basename($this->model) . ' updated successfully'));
         } catch (\Exception $e) {
             if ($request->ajax()) {
                 return response()->json([
                     'success' => false,
-                    'message' => 'Error updating ' . class_basename($this->model),
+                    'message' => __('messages.' . class_basename($this->model) . ' updated failed'),
                     'errors' => $e->getMessage()
                 ], 422);
             }
 
-            return back()->withErrors(['error' => 'Error updating ' . class_basename($this->model)])->withInput();
+            return back()->withErrors(['error' => __('messages.' . class_basename($this->model) . ' updated failed')])->withInput();
         }
     }
 
     public function destroy(Request $request, $id)
     {
+        $this->authorize('delete', $this->model);
+
         try {
             $item = $this->model::findOrFail($id);
             $item->delete();
@@ -116,22 +127,22 @@ abstract class BaseController extends Controller
             if ($request->ajax()) {
                 return response()->json([
                     'success' => true,
-                    'message' => class_basename($this->model) . ' deleted successfully'
+                    'message' => __('messages.' . class_basename($this->model) . ' deleted successfully'),
                 ]);
             }
 
             return redirect()->route($this->routePrefix . '.index')
-                ->with('success', class_basename($this->model) . ' deleted successfully');
+                ->with('success', __('messages.' . class_basename($this->model) . ' deleted successfully'));
         } catch (\Exception $e) {
             if ($request->ajax()) {
                 return response()->json([
                     'success' => false,
-                    'message' => 'Error deleting ' . class_basename($this->model),
+                    'message' => __('messages.' . class_basename($this->model) . ' deleted failed'),
                     'errors' => $e->getMessage()
                 ], 422);
             }
 
-            return back()->withErrors(['error' => 'Error deleting ' . class_basename($this->model)]);
+            return back()->withErrors(['error' => __('messages.' . class_basename($this->model) . ' deleted failed')]);
         }
     }
